requirements.txt
langchain>=0.3.7,<0.4.0
langgraph>=0.2.41,<0.3.0
langchain-core>=0.3.8,<0.4.0
langchain-openai>=0.2.7,<0.3.0
langchain-google-genai>=2.0.4,<3.0.0
python-dotenv>=1.0,<2.0
tqdm>=4.66,<5.0
pydantic>=2.5,<3.0
vllm>=0.15.0
huggingface-hub>=0.20.0
transformers>=4.40.0